NLP/exp_1/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
transformers>=4.48.0
datasets
wandb
omegaconf
hydra-core
scikit-learn
matplotlib
numpy
tqdm
